Summer and Transmitter Should be working
Input/Output Criteria:

Summer:
PL0 - PL7 : IN : 8 12-Bit Player Values ex: "100101110111"
PL_ACT : IN : 8-bit value representing which players are active. For example, if Players 1, 3, and 4 are active: "00011010"
CLK : IN : Standard clock
SUM : OUT : 12-bit output of the averaged output of all active players

Transmitter:
CLK : IN : Standard clock
SUM : IN : 12-bit input from the Summer
SCK : OUT : synced clock designed to be connected to the DAC
BUSY : OUT : indicates the beginning of a serial output. To be connected to the ~SYNC pin of the DAC
S_OUT : OUT : the serial data to be transmitted to the DAC